requirements.txt
numpy==1.25.2
pygame==2.5.1
opencv-python==4.8.0.76
Pillow==10.0.1
